chores - add stateDB implicitly in user_save_solutions instead of explicit add in save-function - move user_save_solutions to utils - remove unneccessary test - remove comments
