+++ desdeo/api/models/nimbus.py
@@ -24,7 +24,7 @@ class NIMBUSClassificationRequest(SQLModel):
 
 class NIMBUSSaveRequest(SQLModel):
     """Request model for saving solutions from any method's state."""
-    # ??:
+
     problem_id: int
     session_id: int | None = Field(default=None)
     parent_state_id: int | None = Field(default=None)
